python/requirements.txt
python-geosupport==1.0.5
sqlalchemy
psycopg2-binary
pandas